refactor(tokenizer): extract filters into separate files
Split pos_filters.py into individual files for better organization:
- particle_filter.py: ParticleFilter implementation
- punctuation_filter.py: PunctuationFilter implementation

This improves code organization and makes it easier to add new
filters in the future. All tests still pass (82 total).
